+++ tests/integration/models/reddit/test_subreddit.py
@@ -4,7 +4,12 @@ from json import dumps
 import socket
 import sys
 
-from praw.exceptions import APIException, ClientException, WebSocketException
+from praw.exceptions import (
+    APIException,
+    ClientException,
+    InvalidFlairTemplateID,
+    WebSocketException,
+)
 from praw.models import (
     Comment,
     ModAction,
@@ -77,11 +82,6 @@ class WebsocketMockException:
                     }
                 }
             )
-
-
-def raise_exception(exception):
-    """Raise the specified exception."""
-    raise exception
 
 
 class TestSubreddit(IntegrationTest):
@@ -1008,6 +1008,122 @@ class TestSubredditFlairTemplates(IntegrationTest):
                 text_color="dark",
                 background_color="#00FFFF",
             )
+
+    @mock.patch("time.sleep", return_value=None)
+    def test_update_invalid(self, _):
+        self.reddit.read_only = False
+        with self.recorder.use_cassette(
+            "TestSubredditFlairTemplates.test_update_invalid"
+        ):
+            with pytest.raises(InvalidFlairTemplateID):
+                self.subreddit.flair.templates.update(
+                    "fake id",
+                    "PRAW updated",
+                    css_class="myCSS",
+                    text_color="dark",
+                    background_color="#00FFFF",
+                    fetch=True,
+                )
+
+    @mock.patch("time.sleep", return_value=None)
+    def test_update_fetch(self, _):
+        self.reddit.read_only = False
+        with self.recorder.use_cassette(
+            "TestSubredditFlairTemplates.test_update_fetch"
+        ):
+            template = list(self.subreddit.flair.templates)[0]
+            self.subreddit.flair.templates.update(
+                template["id"],
+                "PRAW updated",
+                css_class="myCSS",
+                text_color="dark",
+                background_color="#00FFFF",
+                fetch=True,
+            )
+
+    @mock.patch("time.sleep", return_value=None)
+    def test_update_fetch_no_css_class(self, _):
+        self.reddit.read_only = False
+        with self.recorder.use_cassette(
+            "TestSubredditFlairTemplates.test_update_fetch_no_css_class"
+        ):
+            template = list(self.subreddit.flair.templates)[0]
+            self.subreddit.flair.templates.update(
+                template["id"],
+                "PRAW updated",
+                text_color="dark",
+                background_color="#00FFFF",
+                fetch=True,
+            )
+
+    @mock.patch("time.sleep", return_value=None)
+    def test_update_fetch_no_text(self, _):
+        self.reddit.read_only = False
+        with self.recorder.use_cassette(
+            "TestSubredditFlairTemplates.test_update_fetch_no_text"
+        ):
+            template = list(self.subreddit.flair.templates)[0]
+            self.subreddit.flair.templates.update(
+                template["id"],
+                css_class="myCSS",
+                text_color="dark",
+                background_color="#00FFFF",
+                fetch=True,
+            )
+
+    @mock.patch("time.sleep", return_value=None)
+    def test_update_fetch_no_text_or_css_class(self, _):
+        self.reddit.read_only = False
+        with self.recorder.use_cassette(
+            "TestSubredditFlairTemplates.test_update_fetch_"
+            "no_text_or_css_class"
+        ):
+            template = list(self.subreddit.flair.templates)[0]
+            self.subreddit.flair.templates.update(
+                template["id"],
+                text_color="dark",
+                background_color="#00FFFF",
+                fetch=True,
+            )
+
+    @mock.patch("time.sleep", return_value=None)
+    def test_update_fetch_only(self, _):
+        self.reddit.read_only = False
+        with self.recorder.use_cassette(
+            "TestSubredditFlairTemplates.test_update_fetch_only"
+        ):
+            template = list(self.subreddit.flair.templates)[0]
+            self.subreddit.flair.templates.update(
+                template["id"], fetch=True,
+            )
+            newtemplate = list(
+                filter(
+                    lambda _template: _template["id"] == template["id"],
+                    self.subreddit.flair.templates,
+                )
+            )[0]
+            assert newtemplate == template
+
+    @mock.patch("time.sleep", return_value=None)
+    def test_update_false(self, _):
+        self.reddit.read_only = False
+        with self.recorder.use_cassette(
+            "TestSubredditFlairTemplates.test_update_false"
+        ):
+            template = list(self.subreddit.flair.templates)[0]
+            self.subreddit.flair.templates.update(
+                template["id"], text_editable=True, fetch=True,
+            )
+            self.subreddit.flair.templates.update(
+                template["id"], text_editable=False, fetch=True,
+            )
+            newtemplate = list(
+                filter(
+                    lambda _template: _template["id"] == template["id"],
+                    self.subreddit.flair.templates,
+                )
+            )[0]
+            assert newtemplate["text_editable"] is False
 
 
 class TestSubredditLinkFlairTemplates(IntegrationTest):