+++ tests/unit/test_exceptions.py
@@ -3,6 +3,7 @@ from praw.exceptions import (
     APIException,
     ClientException,
     DuplicateReplaceException,
+    InvalidFlairTemplateID,
     InvalidURL,
     InvalidImplicitAuth,
     MissingRequiredAttributeException,
@@ -51,6 +52,18 @@ class TestDuplicateReplaceException:
             str(DuplicateReplaceException())
             == "A duplicate comment has been detected. Are you attempting to "
             "call ``replace_more_comments`` more than once?"
+        )
+
+
+class TestInvalidFlairTemplateID:
+    def test_inheritance(self):
+        assert isinstance(InvalidFlairTemplateID(None), ClientException)
+
+    def test_str(self):
+        assert (
+            str(InvalidFlairTemplateID("123"))
+            == "The flair template id ``123`` is invalid. If you are "
+            "trying to create a flair, please use the ``add`` method."
         )