+++ CHANGES.rst
@@ -20,14 +20,29 @@ Unreleased
 * :meth:`~.Emoji.update` supports modifying an emoji's permissions.
 * :meth:`~.SubredditEmoji.add` now supports optionally passing booleans to set
   an emoji's permissions upon upload.
+* Methods :meth:`.SubredditLinkFlairTemplates.update` and
+  :meth:`.SubredditRedditorFlairTemplates.update` contain a new parameter,
+  ``fetch``, that toggles the automatic fetching of existing data from Reddit.
+  It is set to True by default.
+* Values in methods :meth:`.SubredditLinkFlairTemplates.update` and
+  :meth:`.SubredditRedditorFlairTemplates.update` that are left as the
+  defaults will no longer be over-written if the ``fetch`` parameter is set to
+  ``True``, but will fill in existing values for the flair template.
+* The parameter ``text`` for methods
+  :meth:`.SubredditLinkFlairTemplates.update` and
+  :meth:`.SubredditRedditorFlairTemplates.update` is no longer required.
 
 **Removed**
 
 * Converting :class:`.APIException` to string will no longer escape unicode
   characters.
-* Module ``praw.models.modaction`` no longer exists. Pleae use the module
-  ``praw.models.mod_action`` module, or directly import ``ModAction``
+* Module ``praw.models.modaction`` no longer exists. Please use the module
+  ``praw.models.mod_action``, or directly import ``ModAction``
   from ``praw.models``.
+* Methods :meth:`.SubredditLinkFlairTemplates.update` and
+  :meth:`.SubredditRedditorFlairTemplates.update` will no longer
+  create flairs that are using an invalid template id, but instead throw a
+  :class:`.InvalidFlairTemplateID`.
 
 6.5.1 (2020/01/07)
 ------------------